KTO-1201 valintaperuste käyttöoikeus tyypille
